Move shared search-result components to metabase/common/components
SearchResult, SearchResultLink, and InfoText are presentational
components reused by the global nav search bar (and via it, dashboard
LinkViz), the command palette, and entity pickers. Move the cluster
into metabase/common/components so these non-search consumers no longer
import from the search module.
