+++ frontend/src/metabase/common/components/InfoText/InfoText.tsx
