+++ frontend/src/metabase/common/components/InfoText/InfoTextEditedInfo.styled.tsx
