+++ frontend/src/metabase/common/components/InfoText/get-info-text.tsx
