+++ frontend/src/metabase/common/components/InfoText/index.ts
