+++ frontend/src/metabase/common/components/SearchResult/SearchResult.styled.tsx
