+++ frontend/src/metabase/common/components/SearchResult/SearchResult.tsx
