+++ frontend/src/metabase/common/components/SearchResult/components/DefaultIcon.tsx
