+++ frontend/src/metabase/common/components/SearchResult/components/ItemIcon.tsx
