+++ frontend/src/metabase/common/components/SearchResult/components/constants.ts
