+++ frontend/src/metabase/common/components/SearchResult/components/index.ts
