+++ frontend/src/metabase/common/components/SearchResult/index.ts
