+++ frontend/src/metabase/common/components/SearchResult/tests/SearchResult-Collections.unit.spec.tsx
@@ -10,9 +10,9 @@ import {
   renderWithProviders,
   screen,
 } from "__support__/ui";
+import { SearchResult } from "metabase/common/components/SearchResult";
 import { reinitialize } from "metabase/plugins";
 import { createMockState } from "metabase/redux/store/mocks";
-import { SearchResult } from "metabase/search/components/SearchResult";
 import type {
   SearchResult as ApiSearchResult,
   TokenFeatures,
