+++ frontend/src/metabase/common/components/SearchResult/tests/SearchResult-Tables.unit.spec.tsx
@@ -7,8 +7,8 @@ import {
   setupUsersEndpoints,
 } from "__support__/server-mocks";
 import { renderWithProviders, screen } from "__support__/ui";
-import { SearchResult } from "metabase/search/components/SearchResult";
-import { createWrappedSearchResult } from "metabase/search/components/SearchResult/tests/util";
+import { SearchResult } from "metabase/common/components/SearchResult";
+import { createWrappedSearchResult } from "metabase/common/components/SearchResult/tests/util";
 import type { InitialSyncStatus } from "metabase-types/api";
 import {
   createMockDatabase,
