+++ frontend/src/metabase/common/components/SearchResult/tests/util.ts
