+++ frontend/src/metabase/common/components/SearchResultLink/SearchResultLink.styled.tsx
