+++ frontend/src/metabase/common/components/SearchResultLink/SearchResultLink.tsx
