+++ frontend/src/metabase/common/components/SearchResultLink/index.ts
