+++ frontend/src/metabase/nav/components/search/SearchResults/SearchResults.tsx
@@ -5,6 +5,7 @@ import { t } from "ttag";
 
 import { skipToken, useSearchQuery } from "metabase/api";
 import { EmptyState } from "metabase/common/components/EmptyState";
+import { SearchResult } from "metabase/common/components/SearchResult/SearchResult";
 import { useListKeyboardNavigation } from "metabase/common/hooks/use-list-keyboard-navigation";
 import {
   EmptyStateContainer,
@@ -13,7 +14,6 @@ import {
   SearchResultsList,
 } from "metabase/nav/components/search/SearchResults/SearchResults.styled";
 import { useDispatch } from "metabase/redux";
-import { SearchResult } from "metabase/search/components/SearchResult/SearchResult";
 import { Loader } from "metabase/ui";
 import { modelToUrl } from "metabase/urls";
 import {
